Force requires on project and initials
