+++ src/projects/new/NewProjectForm.tsx
@@ -1,5 +1,8 @@
+import Button from 'react-bootstrap/Button'
+import {useNavigate} from 'react-router-dom'
 import Form from 'react-bootstrap/Form'
 import ProjectState from './projectState'
+import APIUtil from '../../utilities/APIutils'
 
 interface stateProps {
     state: ProjectState
@@ -7,6 +10,18 @@ interface stateProps {
 }
 
 function NewProjectForm({state, setState}: stateProps) {
+    let navigate = useNavigate();
+    const api = new APIUtil()
+
+    const onCancel = ()=> {
+        const  path = '/projects'
+        navigate(path)
+    }
+
+    const onSubmit = () => {
+        api.createProject(state)
+    }
+
     return (
         <div className='new-project-form'>
             <Form>
@@ -19,6 +34,7 @@ function NewProjectForm({state, setState}: stateProps) {
                             value={state.name}
                             onChange={(e)=>setState({...state, name: e.target.value})}
                             placeholder='Project name'
+                            required
                         />
                     </Form.Group>
                     <Form.Group className='mb-3' controlId='formGroupText'>
@@ -29,6 +45,7 @@ function NewProjectForm({state, setState}: stateProps) {
                             value={state.initials}
                             onChange={e=>setState({...state, initials: e.target.value})}
                             placeholder='PBJ'
+                            required
                         />
                     </Form.Group>
                     <Form.Group controlId="formFile" className="mb-3">
@@ -51,6 +68,11 @@ function NewProjectForm({state, setState}: stateProps) {
                             }}
                         />
                     </Form.Group>
+                </div>
+                <div className='new-project-buttons'>
+                    <Button onClick={onCancel} className='new-project-button-cancel'>Back</Button>
+                        <div className='space'></div>
+                    <Button type='submit' className='new-project-button-submit' onClick={onSubmit}>Done</Button>
                 </div>
             </Form>
         </div>
